+++ OCKCatalog/OCKCatalog/AppDelegate.swift
@@ -34,6 +34,12 @@ import UIKit
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
+    let store: OCKStore = {
+        let store = OCKStore(name: "carekit-catalog")
+        store.fillWithDummyData()
+        return store
+    }()
+
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         return true
     }
