+++ OCKSample/OCKSample/AppDelegate.swift
@@ -27,10 +27,21 @@ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+
+import CareKit
+import Contacts
 import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    // Manages synchronization of a CoreData store
+    lazy var synchronizedStoreManager: OCKSynchronizedStoreManager = {
+        let store = OCKStore(name: "SampleAppStore")
+        store.populateSampleData()
+        let manager = OCKSynchronizedStoreManager(wrapping: store)
+        return manager
+    }()
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         return true
@@ -38,5 +49,82 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
         return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
+    }
+}
+
+private extension OCKStore {
+
+    // Adds tasks and contacts into the store
+    func populateSampleData() {
+
+        let thisMorning = Calendar.current.startOfDay(for: Date())
+        let aFewDaysAgo = Calendar.current.date(byAdding: .day, value: -4, to: thisMorning)!
+        let beforeBreakfast = Calendar.current.date(byAdding: .hour, value: 8, to: aFewDaysAgo)!
+        let afterLunch = Calendar.current.date(byAdding: .hour, value: 14, to: aFewDaysAgo)!
+
+        let schedule = OCKSchedule(composing: [
+            OCKScheduleElement(start: beforeBreakfast, end: nil,
+                               interval: DateComponents(day: 1)),
+
+            OCKScheduleElement(start: afterLunch, end: nil,
+                               interval: DateComponents(day: 2))
+        ])
+
+        var doxylamine = OCKTask(id: "doxylamine", title: "Take Doxylamine",
+                             carePlanID: nil, schedule: schedule)
+        doxylamine.instructions = "Take 25mg of doxylamine when you experience nausea."
+
+        let nauseaSchedule = OCKSchedule(composing: [
+            OCKScheduleElement(start: beforeBreakfast, end: nil, interval: DateComponents(day: 1),
+                               text: "Anytime throughout the day", targetValues: [], duration: .allDay)
+            ])
+
+        var nausea = OCKTask(id: "nausea", title: "Track your nausea",
+                             carePlanID: nil, schedule: nauseaSchedule)
+        nausea.impactsAdherence = false
+        nausea.instructions = "Tap the button below anytime you experience nausea."
+
+        let kegelSchedule = OCKSchedule(composing: [OCKScheduleElement(start: beforeBreakfast, end: nil, interval: DateComponents(day: 2))])
+        var kegels = OCKTask(id: "kegels", title: "Kegel Exercises", carePlanID: nil, schedule: kegelSchedule)
+        kegels.impactsAdherence = true
+        kegels.instructions = "Perform kegel exercies"
+
+        addTasks([nausea, doxylamine, kegels], callbackQueue: .main, completion: nil)
+
+        var contact1 = OCKContact(id: "jane", givenName: "Jane",
+                                  familyName: "Daniels", carePlanID: nil)
+        contact1.asset = "JaneDaniels"
+        contact1.title = "Family Practice Doctor"
+        contact1.role = "Dr. Daniels is a family practice doctor with 8 years of experience."
+        contact1.emailAddresses = [OCKLabeledValue(label: CNLabelEmailiCloud, value: "janedaniels@icloud.com")]
+        contact1.phoneNumbers = [OCKLabeledValue(label: CNLabelWork, value: "(324) 555-7415")]
+        contact1.messagingNumbers = [OCKLabeledValue(label: CNLabelWork, value: "(324) 555-7415")]
+
+        contact1.address = {
+            let address = OCKPostalAddress()
+            address.street = "2598 Reposa Way"
+            address.city = "San Francisco"
+            address.state = "CA"
+            address.postalCode = "94127"
+            return address
+        }()
+
+        var contact2 = OCKContact(id: "matthew", givenName: "Matthew",
+                                  familyName: "Reiff", carePlanID: nil)
+        contact2.asset = "MatthewReiff"
+        contact2.title = "OBGYN"
+        contact2.role = "Dr. Reiff is an OBGYN with 13 years of experience."
+        contact2.phoneNumbers = [OCKLabeledValue(label: CNLabelWork, value: "(324) 555-7415")]
+        contact2.messagingNumbers = [OCKLabeledValue(label: CNLabelWork, value: "(324) 555-7415")]
+        contact2.address = {
+            let address = OCKPostalAddress()
+            address.street = "396 El Verano Way"
+            address.city = "San Francisco"
+            address.state = "CA"
+            address.postalCode = "94127"
+            return address
+        }()
+
+        addContacts([contact1, contact2])
     }
 }
